[docs] Add docstrings to coordinates
And move grpc_utils function bodies to source file
